Set pkgrel in python packages of installed version
Every so often Archlinux updates the packaged python version leaving
packages built against the previous latest version obsolete and
non-functional when installed.

This change sets a pkgrel value based on the currently installed system python
version so these packages can be re-built overnight with the rest when
the latest detected python version on the system changes.