Changes and improvemnets to repo_sign
add repoLockfileWait to wait for an add-repo lockfile to go away instead of letting it error on its own.

Add --force-repo-add to force adding all found packages of a repo (with --new and --sign)

Split add-repo from signThing function to repoAddThing function

Have repoAddThing remove a signature file if it fails to add a the package
to the repo. So that future runs won't miss it.

Remove signature of single